Week 3 Assignment
Hi Matt,

I realize this is not quite right, but it is about as far as I could
get by tonight (Saturday).

I had some unexplained trouble trying to float the <article> section
next to the <aside> section, but for some reason when I put each in an
extra div it worked immediately, so that explains the extra container.

The “sign up now!” button was also not settling in nicely to the
footer, which is why I did the “position: relative;” with it, which I
know is probably not the best way.